requirements.txt
torch
diffusers
transformers
accelerate
numpy
opencv-python
Pillow
tqdm